+++ app/__init__.py
@@ -3,8 +3,8 @@ from flask import Flask
 
 app = Flask(__name__)
 
-app.config['VERSION'] = '0.34.1'
-app.config['VERSION_DATE'] = '26 August 2026'
+app.config['VERSION'] = '0.37.0'
+app.config['VERSION_DATE'] = '28 August 2026'
 app.config['VERSION_DATE_ISO'] = datetime.strptime(app.config['VERSION_DATE'], '%d %B %Y').strftime('%Y-%m-%d')
 
 @app.context_processor